refactor out resume save logic awawy from gemin agent
