+++ src/commands_automator/llm/llm_agents/gemini/gemini_mcp_client.py
@@ -13,6 +13,7 @@ from google.genai.types import GenerateContentResponse
 from PIL import Image
 import io
 
+from commands_automator.llm.llm_agents.agent_services.resume_refiner_service import ResumeRefinerService
 from commands_automator.llm.llm_agents.agent_services.resume_saver_service import ResumeSaverService
 
 
@@ -35,8 +36,9 @@ class SmartMCPClient:
             "max_output_tokens": 8192,  # Limit response length
             "response_mime_type": "application/json",  # Output as plain text
         }
+
         self.resume_chat._config = config
-        self.resume_saver_service: ResumeSaverService = ResumeSaverService()
+        self.resume_refiner_service = ResumeRefinerService(self.resume_chat)
 
     async def decide_tool_usage(self, query, available_tools, session):
         """
@@ -227,15 +229,7 @@ Query: {query}
     
     def refine_resume(self, tool_result, output_file_path):
         if self.api_key:
-            try:
-                resume_data_dict = json.loads(tool_result)
-            except ValueError as ve:
-                logging.error(f"error with json structure of tool {ve}")
-                return ""
-            resume_text = self.get_refined_resume(resume_data_dict)
-            cover_letter_text = self.get_cover_letter(resume_data_dict)
-            self.save_resume_files(output_file_path, resume_data_dict, resume_text, cover_letter_text)
-            return resume_text + "\n\n" + cover_letter_text
+            return self.resume_refiner_service.refine_resume(tool_result, output_file_path)
         else:
             return tool_result
 
@@ -257,14 +251,6 @@ Query: {query}
         else:
             return "No suitable tool found and Gemini API key not provided."
 
-    def save_resume_files(self, output_file_path, resume_data_dict, resume_text, cover_letter_text):
-        resume_highlighted_sections = self.convert_none_to_empty_string(resume_data_dict.get('resume_highlighted_sections', ''))
-        applicant_name = resume_data_dict.get('applicant_name', '')
-        resume_file_name = self.resume_saver_service.get_resume_file_name(resume_text, applicant_name)
-        self.resume_saver_service.save_resume(resume_text, output_file_path, applicant_name, resume_file_name, resume_highlighted_sections)
-        if (cover_letter_text != ''):
-            self.resume_saver_service.save_cover_letter(cover_letter_text, output_file_path, applicant_name, resume_file_name)
-
     async def get_available_tools(self, session):
         try:
             tools_response = await session.list_tools()
@@ -275,30 +261,6 @@ Query: {query}
             available_tools = []
         return available_tools
 
-    def get_cover_letter(self, resume_data_dict):
-        cover_letter_guidelines = resume_data_dict.get('cover_letter_guidelines', '')
-        cover_letter_text = ''
-        if cover_letter_guidelines is not None:
-            gemini_response = self.resume_chat.send_message(cover_letter_guidelines)
-            cover_letter_text = gemini_response._get_text()
-        return cover_letter_text
-
-    def get_refined_resume(self, resume_data_dict):
-        prompt = self.format_prompts_for_resume(resume_data_dict)
-        self.resume_chat._config["response_mime_type"] = "text/plain"
-        gemini_response = self.resume_chat.send_message(prompt)
-        resume_text = gemini_response._get_text()
-        return resume_text
-        
-    def format_prompts_for_resume(self, resume_data_dict):       
-        general_guidleines = resume_data_dict.get('general_guidelines', '')
-        resume = resume_data_dict.get('resume', '')        
-        jobs_desc = self.convert_none_to_empty_string(resume_data_dict.get('job_description', ''))
-     
-        prompt = f"""You have finished using the mcp tool. Now output text according to the following guidelines.\n\n
-                    {general_guidleines}\n\n{resume}\n\n\n"""
-        prompt += jobs_desc
-        return prompt 
 
     def convert_none_to_empty_string(self, text):
         if text == None:
